Modello di costo LLM calibrato (equazione stimata dal testo sorgente)
- estimate: per il motore openai il costo e' stimato dal testo sorgente con
  l'equazione calibrata su una traduzione reale (Mercury-2.5, 15 pagine):
  token_in=(chars/4)*13.3, token_out=token_in*1.09, costo=token/1e6*prezzo
- Esito: ~0.0054 USD/pagina (misurato 0.00537; la stima naive era 13x in meno)
- COST_CENTS_PER_PAGE_* come override commerciale; cost_cents ora frazionario
- Env: LLM_PRICE_*, LLM_OVERHEAD_FACTOR, LLM_OUTPUT_RATIO, CHARS_PER_TOKEN,
  ESTIMATE_SAMPLE_PAGES; README con l'equazione; frontend mostra la valuta
- Test: stima costo LLM (60 test)

+++ app/models.py
@@ -159,7 +159,7 @@ class EstimateOut(BaseModel):
     pages_to_translate: int
     ms_per_page: int
     estimated_seconds: int
-    cost_cents: int
+    cost_cents: float
     currency: str = "EUR"
     note: str | None = None
